tests/submission/W3C/domhtml/HTMLIFrameElement02.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - IFRAME</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-iframe-frameborder" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<iframe frameborder="1" src="/support/html/sample.html" height='80' width="550">IFRAME1</iframe>


<script>
var domtest = async_test("The frameBorder IDL attribute of the frame element must reflect the element's frameborder content attribute.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vframeborder;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("iframe");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vframeborder = testNode.frameBorder;

      assert_equals(vframeborder,"1");
       
    domtest.done();
}));

</script>
<p>The frameBorder IDL attribute of the frame element must reflect the element's frameborder content attribute.</p>
<div id="log"></div>
</body></html>
